--- .env.sh
@@ -1,3 +1,0 @@
-# Sets up a few convenient aliases and env vars.
-alias g=./gradlew
-export GRADLE_USER_HOME=$HOME/.gradle
