+++ elastiknn-benchmarks/deploy/datasets-workflow.yaml
@@ -126,7 +126,7 @@ spec:
         resources:
           requests:
             cpu: 1
-            memory: 2Gi
+            memory: 8Gi
           limits:
             cpu: 1
-            memory: 2Gi
+            memory: 8Gi
