+++ elastiknn-benchmarks/python/requirements.txt
@@ -6,3 +6,5 @@ elastiknn-client
 h5py==2.10.0
 pareto==1.1.1-3
 pandas==1.1.2
+scikit-learn==0.21.0
+tabulate==0.8.7
