+++ elastiknn-benchmarks/src/main/resources/logback.xml
@@ -2,7 +2,7 @@
 <configuration>
     <appender name="STDOUT" class="ch.qos.logback.core.ConsoleAppender">
         <layout class="ch.qos.logback.classic.PatternLayout">
-            <Pattern>[%d{HH:mm:ss.SSS}] [%thread] [%-5level] [%logger{140}] %msg%n</Pattern>
+            <Pattern>[%d{HH:mm:ss.SSS}] [%thread] [%level] [%logger{140}] %msg%n</Pattern>
         </layout>
     </appender>
     <logger name="com.sksamuel.elastic4s" level="WARN"/>
