+++ elastiknn-benchmarks/src/main/scala/com/klibisz/elastiknn/benchmarks/SearchClient.scala
@@ -91,7 +91,7 @@ object SearchClient {
                 case (query, i) =>
                   for {
                     (dur, res) <- ZIO.fromFuture(_ => client.nearestNeighbors(index, query, k, "id")).timed
-                    _ <- if (i % 100 == 0) log.debug(s"Completed query [$i] in [$index] in [${dur.toMillis}] ms") else ZIO.succeed(())
+                    _ <- if (i % 100 == 0) log.info(s"Completed query [$i] in [$index] in [${dur.toMillis}] ms") else ZIO.succeed(())
                   } yield QueryResult(res.result.hits.hits.map(_.score), res.result.took)
               }
