+++ elastiknn-benchmarks/src/main/scala/com/klibisz/elastiknn/benchmarks/package.scala
@@ -101,98 +101,6 @@ package object benchmarks {
     }
   }
 
-  object Experiment {
-
-    private val vecName: String = "vec"
-
-    /**
-      * Alias to gridsearch method for multiple datasets.
-      */
-    def gridsearch(datasets: Dataset*): Seq[Experiment] = datasets.flatMap(gridsearch)
-
-    /**
-      * Returns reasonable Experiments for doing a gridsearch over parameters for the given dataset.
-      */
-    def gridsearch(dataset: Dataset): Seq[Experiment] = dataset match {
-
-      case Dataset.AnnbMnist | Dataset.AnnbFashionMnist =>
-        Seq(
-          Experiment(
-            dataset,
-            Mapping.DenseFloat(dataset.dims),
-            Seq(Query(NearestNeighborsQuery.Exact(vecName, Similarity.L2), 100))
-          )
-        ) ++ {
-          for {
-            tables <- Seq(50, 75, 100)
-            hashesPerTable <- Seq(2, 3, 4)
-            width <- 5 to 8
-          } yield
-            Experiment(
-              dataset,
-              Mapping.L2Lsh(dataset.dims, L = tables, k = hashesPerTable, w = width),
-              for {
-                candidates <- Seq(1000, 5000)
-                probes <- Seq(0, 3, 6, 9)
-              } yield Query(NearestNeighborsQuery.L2Lsh(vecName, candidates, probes), 100)
-            )
-        }
-
-      case Dataset.AnnbSift =>
-        Seq(
-          Experiment(
-            dataset,
-            Mapping.DenseFloat(dataset.dims),
-            Seq(Query(NearestNeighborsQuery.Exact(vecName, Similarity.L2), 100))
-          )) ++ {
-          for {
-            tables <- Seq(50, 75, 100)
-            hashesPerTable <- Seq(2, 3, 4)
-            width <- Seq(1, 2, 3)
-          } yield
-            Experiment(
-              dataset,
-              Mapping.L2Lsh(dataset.dims, L = tables, k = hashesPerTable, w = width),
-              for {
-                candidates <- Seq(1000, 5000, 10000)
-                probes <- 0 to math.pow(hashesPerTable, 3).toInt.min(9) by 3
-              } yield Query(NearestNeighborsQuery.L2Lsh(vecName, candidates, probes), 100)
-            )
-        }
-
-      case Dataset.AnnbGlove100 =>
-        val exact = Seq(
-          Experiment(dataset, Mapping.DenseFloat(dataset.dims), Seq(Query(NearestNeighborsQuery.Exact(vecName, Similarity.Angular), 100)))
-        )
-        val projections = for {
-          tables <- Seq(50, 100, 125)
-          hashesPerTable <- Seq(6, 9)
-        } yield
-          Experiment(
-            dataset,
-            Mapping.AngularLsh(dataset.dims, tables, hashesPerTable),
-            for {
-              candidates <- Seq(1000, 5000)
-            } yield Query(NearestNeighborsQuery.AngularLsh(vecName, candidates), 100)
-          )
-        val permutations = for {
-          k <- Seq(10, 25, 50, 75)
-          rep <- Seq(false)
-        } yield
-          Experiment(
-            dataset,
-            Mapping.PermutationLsh(dataset.dims, k, repeating = rep),
-            for {
-              candidates <- Seq(1000, 5000, 10000)
-            } yield Query(NearestNeighborsQuery.PermutationLsh(vecName, Similarity.Angular, candidates), 100)
-          )
-        exact ++ projections ++ permutations
-
-      case _ => Seq.empty
-    }
-
-  }
-
   object codecs {
     private implicit val mappingCodec: Codec[Mapping] = ElasticsearchCodec.mapping
     private implicit val nnqCodec: Codec[NearestNeighborsQuery] = ElasticsearchCodec.nearestNeighborsQuery
